+++ src/ecotally/__init__.py
@@ -1,6 +1,6 @@
 """Small, transparent tools for community ecology."""
 
-__version__ = "0.12.0"
+__version__ = "0.13.0"
 
 from .beta import PairwiseResult, compare_communities
 from .diversity import DiversityResult, calculate_diversity
@@ -8,6 +8,7 @@ from .estimation import RichnessEstimate, estimate_richness, rarefaction_curve
 from .summary import DatasetSummary, summarize_dataset, summarize_species
 from .quality import QualityIssue, audit_communities
 from .uncertainty import bootstrap_diversity
+from .functional import FunctionalResult, calculate_functional_diversity
 
 __all__ = [
     "DiversityResult",
@@ -23,4 +24,6 @@ __all__ = [
     "QualityIssue",
     "audit_communities",
     "bootstrap_diversity",
+    "FunctionalResult",
+    "calculate_functional_diversity",
 ]